Revert "Added vanilla LSTM and LSTM with peepholes oneDNN fp32 kernel (#30661)"
This reverts commit d834f4e6e86099e5a6d194e5b888ff03cc92d32c.
