+++ python/paddle/fluid/tests/unittests/mkldnn/test_fusion_gru_mkldnn_op.py
@@ -75,6 +75,4 @@ class TestFusionGRUMKLDNNOpBS1(TestFusionGRUOp):
 
 
 if __name__ == "__main__":
-    from paddle import enable_static
-    enable_static()
     unittest.main()
