+++ python/paddle/fluid/tests/unittests/test_fusion_gru_op.py
@@ -144,6 +144,4 @@ class TestFusionGRUOpBS1(TestFusionGRUOp):
 
 
 if __name__ == "__main__":
-    from paddle import enable_static
-    enable_static()
     unittest.main()
