+++ python/paddle/fluid/tests/unittests/test_fusion_lstm_op.py
@@ -58,7 +58,6 @@ class TestFusionLSTMOp(OpTest):
         self.act_gate = 'sigmoid'
         self.act_cell = 'tanh'
         self.act_cand = 'tanh'
-        self.use_mkldnn = False
         self.set_conf()
 
         T = sum(self.lod[0])
@@ -111,8 +110,7 @@ class TestFusionLSTMOp(OpTest):
             'is_reverse': self.is_reverse,
             'gate_activation': self.act_gate,
             'cell_activation': self.act_cell,
-            'candidate_activation': self.act_cand,
-            'use_mkldnn': self.use_mkldnn
+            'candidate_activation': self.act_cand
         }
 
     def test_check_output(self):
@@ -193,6 +191,4 @@ class TestFusionLSTMOpPeepholesBS1(TestFusionLSTMOp):
 
 
 if __name__ == '__main__':
-    from paddle import enable_static
-    enable_static()
     unittest.main()
