+++ python/paddle/fluid/tests/unittests/white_list/no_check_set_white_list.py
@@ -29,5 +29,4 @@ no_check_set_white_list = [
     'update_loss_scaling',
     'cudnn_lstm',
     'rnn',
-    'fusion_lstm',
 ]
