+++ tools/static_mode_white_list.py
@@ -601,7 +601,6 @@ STATIC_MODE_TESTING_LIST = [
     'test_bilinear_interp_mkldnn_op',
     'test_fusion_gru_int8_mkldnn_op',
     'test_fusion_gru_mkldnn_op',
-    'test_fusion_lstm_mkldnn_op',
     'test_gaussian_random_mkldnn_op',
     'test_lrn_mkldnn_op',
     'test_matmul_mkldnn_op',
